pyproject.toml
[project]
name = "pipeline"
version = "0.1.0"
description = "Multi-source streaming data pipeline for insurance claims processing"
requires-python = ">=3.12"
dependencies = [
    "confluent-kafka>=2.6.0",
    "azure-eventhub>=5.11.0",
    "structlog>=24.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
